+++ R/fimport.R
@@ -1,0 +1,10 @@
+
+# Import function ---------------------------------------------------------
+
+
+
+fimport <- function(path, engine = NULL) {
+  
+  
+  
+}